menmbahkan algoritma CRUD pada dokter dan juga pasien, dokter dapat memeriksa dan pasien dapat periksa

+++ database/migrations/2025_03_18_045233_create_obats_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama_obat');
             $table->string('kemasan');
-            $table->bigInteger('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
